+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixes
+
+- Auth & Human Verification: Closing Human Verification screen on sign up goes back to plans.
+
 ## [3.0.0]
 
 ### Changes